refactor(uta-service): relocate simulator routes from Alice to UTA
Step 6 staging — `/api/simulator/*` mounts move from Alice to UTA so
the MockBroker god-view runs in the same process as the MockBroker
instances it manipulates. Alice's BFF proxies the prefix to UTA the
same way it already proxies `/api/trading/*`.

  Was:   Alice plugin.ts → createSimulatorRoutes(ctx)
         (touched ctx.utaManager.get(id).broker directly)
  Now:   UTA main.ts     → createSimulatorRoutes(tradingCtx)
         Alice plugin.ts → createTradingProxyRoutes (catches both
                          /api/trading/* and /api/simulator/*)

Move was via git mv so the 9 simulator routes + the spec keep their
history. Imports flipped from relative `../../core/...` /
`../../domain/...` to the `@/` alias which the services/uta tsconfig
remaps to the main repo's src/ (same trick `routes-trading.ts` uses).

This was the last consumer of `ctx.utaManager` that needed an in-
process Alice reference for the trading domain — telegram and the AI
tool layer now stand alone as the remaining migration work.

Verified: tsc clean, pnpm test 1738/1738 pass, `curl :47331/api/
simulator/utas` returns 200 via BFF (the empty list comes through
unchanged from UTA).
