+++ services/uta/src/http/routes-simulator.ts
@@ -17,9 +17,9 @@
 import { Hono } from 'hono'
 import type { Context } from 'hono'
 import { z } from 'zod'
-import type { EngineContext } from '../../core/types.js'
-import { MockBroker } from '../../domain/trading/brokers/mock/MockBroker.js'
-import { SEC_TYPES, type SecType } from '../../domain/trading/contract-discipline.js'
+import type { EngineContext } from '@/core/types.js'
+import { MockBroker } from '@/domain/trading/brokers/mock/MockBroker.js'
+import { SEC_TYPES, type SecType } from '@/domain/trading/contract-discipline.js'
 
 // ==================== Schemas ====================
