+++ services/uta/src/http/simulator.spec.ts
@@ -6,9 +6,9 @@
  */
 
 import { describe, it, expect, beforeEach } from 'vitest'
-import { createSimulatorRoutes } from './simulator.js'
-import { MockBroker } from '../../domain/trading/brokers/mock/MockBroker.js'
-import type { EngineContext } from '../../core/types.js'
+import { createSimulatorRoutes } from './routes-simulator.js'
+import { MockBroker } from '@/domain/trading/brokers/mock/MockBroker.js'
+import type { EngineContext } from '@/core/types.js'
 
 interface FakeUTA {
   id: string
